backend/requirement.txt
Flask
Flask-Cors
Flask-MySQLdb
PyJWT
Flask-RESTful
Flask-SQLAlchemy
WeasyPrint
